foerr: Move MainMenu to a subdir
Create a subdirectory for all things related to main menu. For now it's
only one class, but more will be added.
