+++ src/hud/main_menu/main_menu.cpp
@@ -4,11 +4,11 @@
 
 #include "main_menu.hpp"
 
-#include "../settings/settings_manager.hpp"
-#include "../util/i18n.hpp"
+#include "../../settings/settings_manager.hpp"
+#include "../../util/i18n.hpp"
+#include "../loading_screen.hpp"
+#include "../log.hpp"
 #include "git_version.h"
-#include "loading_screen.hpp"
-#include "log.hpp"
 
 MainMenu::MainMenu(ResourceManager& resMgr, CursorManager& cursorMgr, sf::RenderWindow& window, Campaign& campaign,
 				   GameState& gameState, PipBuck& pipBuck) :
