+++ src/hud/main_menu/main_menu.hpp
@@ -10,15 +10,15 @@
 #include <SFML/Graphics/RenderTarget.hpp>
 #include <SFML/Graphics/RenderWindow.hpp>
 
-#include "../campaigns/campaign.hpp"
-#include "../resources/sound_resource.hpp"
-#include "../window/cursor_manager.hpp"
-#include "buttons/simple_button.hpp"
-#include "click_manager.hpp"
-#include "configurable_gui_component.hpp"
-#include "hover_manager.hpp"
-#include "pipbuck/pipbuck.hpp"
-#include "text_label.hpp"
+#include "../../campaigns/campaign.hpp"
+#include "../../resources/sound_resource.hpp"
+#include "../../window/cursor_manager.hpp"
+#include "../buttons/simple_button.hpp"
+#include "../click_manager.hpp"
+#include "../configurable_gui_component.hpp"
+#include "../hover_manager.hpp"
+#include "../pipbuck/pipbuck.hpp"
+#include "../text_label.hpp"
 
 class MainMenu : public sf::Drawable, public sf::Transformable, public ConfigurableGuiComponent
 {
